Moved filling scheme and bunch selection to 2nd gen + some cleaning
